added session validation for all url endpoints

+++ flask-frontend/app.py
@@ -16,16 +16,43 @@ Session(app)
 # login routes
 @app.route('/')
 def employer_index():
+    if session.get("type") == "employer":
+        return redirect("/employer/list")
+
+    if session.get("type") == "staff":
+        return redirect("/staff/list")
+
+    if session.get("type") == "admin":
+        return redirect("/admin/list")
+
     return render_template("logins/employer-login.html")
 
 
 @app.route('/admin')
 def admin_index():
+    if session.get("type") == "employer":
+        return redirect("/employer/list")
+
+    if session.get("type") == "staff":
+        return redirect("/staff/list")
+
+    if session.get("type") == "admin":
+        return redirect("/admin/list")
+
     return render_template("logins/admin-login.html")
 
 
 @app.route('/staff')
 def staff_index():
+    if session.get("type") == "employer":
+        return redirect("/employer/list")
+        
+    if session.get("type") == "staff":
+        return redirect("/staff/list")
+
+    if session.get("type") == "admin":
+        return redirect("/admin/list")
+        
     return render_template("logins/staff-login.html")
 
 
@@ -121,30 +148,49 @@ def edit_student_staff(id):
     return redirect("/staff")
 
 
-############ TAKE A LOOK!!!!!!!!!!!!!!!!
 @app.route("/admin/list/studentUpdated/", methods=["POST"])
 def edited_student_admin():
-    delete_student_from_api(request.form.get("id"))
-    add_student_to_list(
-        request.form.get("student_id"),
-        request.form.get("first_name"),
-        request.form.get("last_name"),
-        request.form.get("email"),
-        request.form.get("gender"),
-        request.form.get("professor_name"),
-        request.form.get("year_of_graduation"),
-        request.form.get("degree"),
-        request.form.get("projectId"),
-        request.form.get("programming_language")
-    )
-    return redirect ("/admin/list")
+    if request.method == "POST" and session.get("type") == "admin":
+        delete_student_from_api(request.form.get("id"))
+        add_student_to_list(
+            request.form.get("student_id"),
+            request.form.get("first_name"),
+            request.form.get("last_name"),
+            request.form.get("email"),
+            request.form.get("gender"),
+            request.form.get("professor_name"),
+            request.form.get("year_of_graduation"),
+            request.form.get("degree"),
+            request.form.get("projectId"),
+            request.form.get("programming_language")
+        )
+        return redirect ("/admin/list")
+    return redirect("/")
 
-############ TAKE A LOOK!!!!!!!!!!!!!!!!
 @app.route("/staff/list/studentUpdated/", methods=["POST"])
 def edited_student_staff():
-    delete_student_from_api(request.form.get("id"))
-    add_student_to_list(
-        request.form.get("student_id"),
+    if request.method == "POST" and session.get("type") == "staff":
+        delete_student_from_api(request.form.get("id"))
+        add_student_to_list(
+            request.form.get("student_id"),
+            request.form.get("first_name"),
+            request.form.get("last_name"),
+            request.form.get("email"),
+            request.form.get("gender"),
+            request.form.get("professor_name"),
+            request.form.get("year_of_graduation"),
+            request.form.get("degree"),
+            request.form.get("projectId"),
+            request.form.get("programming_language")
+        )
+        return redirect ("/staff/list")
+    return redirect("/")
+    
+@app.route("/staff/list/student/studentadded",methods=["POST"])
+def student_added_from_form_staff():
+    if request.method == "POST" and session.get("type") == "staff":
+        add_student_to_list(
+        request.form.get("student_Id"),
         request.form.get("first_name"),
         request.form.get("last_name"),
         request.form.get("email"),
@@ -154,45 +200,27 @@ def edited_student_staff():
         request.form.get("degree"),
         request.form.get("projectId"),
         request.form.get("programming_language")
-    )
-    return redirect ("/staff/list")
-    
-############ TAKE A LOOK!!!!!!!!!!!!!!!!
-@app.route("/staff/list/student/studentadded",methods=["POST"])
-def student_added_from_form_staff():
+        )
+        return redirect("/staff/list")
+    return redirect("/")
 
-    #call function to post to student api
-    add_student_to_list(
-    request.form.get("student_Id"),
-    request.form.get("first_name"),
-    request.form.get("last_name"),
-    request.form.get("email"),
-    request.form.get("gender"),
-    request.form.get("professor_name"),
-    request.form.get("year_of_graduation"),
-    request.form.get("degree"),
-    request.form.get("projectId"),
-    request.form.get("programming_language")
-    )
-    return redirect("/staff/list")
-
-############ TAKE A LOOK!!!!!!!!!!!!!!!!
 @app.route("/admin/list/student/studentadded",methods=["POST"])
 def student_added_from_form_admin():
-    #call function to post to student api
-    add_student_to_list(
-    request.form.get("student_Id"),
-    request.form.get("first_name"),
-    request.form.get("last_name"),
-    request.form.get("email"),
-    request.form.get("gender"),
-    request.form.get("professor_name"),
-    request.form.get("year_of_graduation"),
-    request.form.get("degree"),
-    request.form.get("projectId"),
-    request.form.get("programming_language")
-    )
-    return redirect("/admin/list")
+    if request.method == "POST" and session.get("type") == "admin":
+        add_student_to_list(
+        request.form.get("student_Id"),
+        request.form.get("first_name"),
+        request.form.get("last_name"),
+        request.form.get("email"),
+        request.form.get("gender"),
+        request.form.get("professor_name"),
+        request.form.get("year_of_graduation"),
+        request.form.get("degree"),
+        request.form.get("projectId"),
+        request.form.get("programming_language")
+        )
+        return redirect("/admin/list")
+    return redirect("/")
 
 @app.route("/admin/pannel")
 def admin_pannel():
@@ -206,102 +234,111 @@ def staff_pannel():
         return render_template("staff-project-degree-pannel.html",DEGREES=get_all_degrees_from_api(),PROJECTS = get_all_projects_from_api())
     return redirect("/staff")
 
-# TAKE A LOOK AT THIS!!!!!!!!!!!!
 @app.route("/admin/pannel/deletedegree",methods=["POST"])
 def delete_a_degree_by_id_route():
-    delete_degree_from_api(request.form.get("id"))
-    return redirect("/admin/pannel")
+    if request.method == "POST" and session.get("type") == "admin":
+        delete_degree_from_api(request.form.get("id"))
+        return redirect("/admin/pannel")
+    return redirect("/")
 
-# TAKE A LOOK AT THIS!!!!!!!!!!!!!!!!!!!!
 @app.route("/admin/pannel/adddegree",methods=["POST"])
 def post_a_degree_route():
-    post_a_degree_to_api(
-        request.form.get("degree_id"),
-        request.form.get("name_degree"),
-        request.form.get("description")
-    )
-    return redirect("/admin/pannel")
+    if request.method == "POST" and session.get("type") == "admin":
+        post_a_degree_to_api(
+            request.form.get("degree_id"),
+            request.form.get("name_degree"),
+            request.form.get("description")
+        )
+        return redirect("/admin/pannel")
+    return redirect("/")
 
-# TAKE A LOOK AT THIS!
 @app.route("/staff/pannel/deletedegree",methods=["POST"])
 def delete_a_degree_by_id_staff_route():
-    delete_degree_from_api(request.form.get("id"))
-    return redirect("/staff/pannel")
+    if request.method == "POST" and session.get("type") == "staff":
+        delete_degree_from_api(request.form.get("id"))
+        return redirect("/staff/pannel")
+    return redirect("/")
 
-# TAKE A LOOK AT THIS!
 @app.route("/staff/pannel/adddegree",methods=["POST"])
 def post_a_degree_staff_route():
-    post_a_degree_to_api(
-        request.form.get("degree_id"),
-        request.form.get("name_degree"),
-        request.form.get("description")
-    )
-    return redirect("/staff/pannel")
+    if request.method == "POST" and session.get("type") == "staff":
+        post_a_degree_to_api(
+            request.form.get("degree_id"),
+            request.form.get("name_degree"),
+            request.form.get("description")
+        )
+        return redirect("/staff/pannel")
+    return redirect("/")
 
-# TAKE A LOOK AT THIS!
 @app.route("/admin/pannel/deleteproject",methods=["POST"])
 def delete_a_project_by_id_route():
-    delete_project_from_api(request.form.get("project_id"))
-    return redirect("/admin/pannel")
+    if request.method == "POST" and session.get("type") == "admin":
+        delete_project_from_api(request.form.get("project_id"))
+        return redirect("/admin/pannel")
+    return redirect("/")
 
-# TAKE A LOOK AT THIS!
 @app.route("/admin/pannel/addproject",methods=["POST"])
 def post_a_project_route():
-    post_a_project_to_api(
-        request.form.get("project_id"),
-        request.form.get("project_name"),
-        request.form.get("project_description")
-    )
-    return redirect("/admin/pannel")
+    if request.method == "POST" and session.get("type") == "admin":
+        post_a_project_to_api(
+            request.form.get("project_id"),
+            request.form.get("project_name"),
+            request.form.get("project_description")
+        )
+        return redirect("/admin/pannel")
+    return redirect("/")
 
-# TAKE A LOOK AT THIS!
 @app.route("/staff/pannel/deleteproject",methods=["POST"])
 def delete_a_project_by_id_route_staff():
-    delete_project_from_api(request.form.get("project_id"))
-    return redirect("/staff/pannel")
+    if request.method == "POST" and session.get("type") == "staff":
+        delete_project_from_api(request.form.get("project_id"))
+        return redirect("/staff/pannel")
+    return redirect("/")
 
-# TAKE A LOOK AT THIS!
 @app.route("/staff/pannel/addproject",methods=["POST"])
 def post_a_project_route_staff():
-    post_a_project_to_api(
-        request.form.get("project_id"),
-        request.form.get("project_name"),
-        request.form.get("project_description")
-    )
-    return redirect("/staff/pannel")
+    if request.method == "POST" and session.get("type") == "staff":
+        post_a_project_to_api(
+            request.form.get("project_id"),
+            request.form.get("project_name"),
+            request.form.get("project_description")
+        )
+        return redirect("/staff/pannel")
 
-# TAKE A LOOK AT THIS!
 @app.route("/admin/pannel/updatedegree",methods=["POST"])
 def update_a_degree_route_admin():
-    put_a_degree_to_api(
-        request.form.get("id"),
-        request.form.get("degree_id"),
-        request.form.get("name_degree"),
-        request.form.get("description")
-    )
-    return redirect("/admin/pannel")
+    if request.method == "POST" and session.get("type") == "admin":
+        put_a_degree_to_api(
+            request.form.get("id"),
+            request.form.get("degree_id"),
+            request.form.get("name_degree"),
+            request.form.get("description")
+        )
+        return redirect("/admin/pannel")
+    return redirect("/")
 
-# TAKE A LOOK AT THIS!
 @app.route("/admin/credentials/deletecredential", methods=["POST"])
 def delete_a_credential_route():
-    delete_credentials_from_api(request.form.get("id"))
-    return redirect("/admin/pannel")
+    if request.method == "POST" and session.get("type") == "admin":
+        delete_credentials_from_api(request.form.get("id"))
+        return redirect("/admin/pannel")
+    return redirect("/")
 
-# TAKE A LOOK AT THIS!
 @app.route("/admin/credentials/addcredential", methods=["POST"])
 def add_credential_route():
-    email = request.form.get("email")
-    password = request.form.get("password")
-    lang = request.form.get("lang")
-    session_type = request.form.get("type")
+    if request.method == "POST" and session.get("type") == "admin":
+        email = request.form.get("email")
+        password = request.form.get("password")
+        lang = request.form.get("lang")
+        session_type = request.form.get("type")
 
-    hashed_pw = bcrypt.hashpw(password.encode('utf-8'), bcrypt.gensalt()).decode()
+        hashed_pw = bcrypt.hashpw(password.encode('utf-8'), bcrypt.gensalt()).decode()
 
-    post_credentials_from_api(email,hashed_pw,lang,session_type)
+        post_credentials_from_api(email,hashed_pw,lang,session_type)
 
-    return redirect("/admin/pannel")
+        return redirect("/admin/pannel")
+    return redirect("/")
 
-# TAKE A LOOK AT THIS!
 @app.route("/admin/login", methods=["POST"])
 def validate_admin_login():
     email = request.form.get("email")
@@ -322,7 +359,6 @@ def validate_admin_login():
                 return redirect("/admin/list")
     return redirect("/admin")
 
-# TAKE A LOOK AT THIS!
 @app.route("/staff/login", methods=["POST"])
 def validate_staff_login():
     email = request.form.get("email")
@@ -343,7 +379,6 @@ def validate_staff_login():
                 return redirect("/staff/list")
     return redirect("/staff")
 
-# TAKE A LOOK AT THIS!
 @app.route("/employer/login", methods=["POST"])
 def validate_employer_login():
     email = request.form.get("email")
